+++ src/GitHub_Updater/API/Bitbucket_API.php
@@ -165,9 +165,9 @@ class Bitbucket_API extends API implements API_Interface {
 
 		/*
 		 * If a branch has been given, use branch.
-		 * If branch is release branch (default) and tags are used, use newest tag.
+		 * If branch is production branch (default) and tags are used, use newest tag.
 		 */
-		if ( $this->type->release_branch !== $this->type->branch || empty( $this->type->tags ) ) {
+		if ( $this->type->production_branch !== $this->type->branch || empty( $this->type->tags ) ) {
 			if ( ! empty( $this->type->enterprise_api ) ) {
 				$endpoint = add_query_arg( 'at', $this->type->branch, $endpoint );
 			} else {