+++ src/GitHub_Updater/Base.php
@@ -232,9 +232,9 @@ class Base {
 	 */
 	public function add_extra_headers() {
 		$ghu_extra_headers = [
-			'RequiresWP'    => 'Requires WP',
-			'ReleaseAsset'  => 'Release Asset',
-			'ReleaseBranch' => 'Release Branch',
+			'RequiresWP'       => 'Requires WP',
+			'ReleaseAsset'     => 'Release Asset',
+			'ProductionBranch' => 'Production Branch',
 		];
 
 		$uri_types = [
@@ -595,9 +595,9 @@ class Base {
 
 		print '<ul id="' . esc_attr( $data['id'] ) . '" style="display:none; width: 100%;">';
 
-		// Disable branch switching to release branch for release assets.
+		// Disable branch switching to production branch for release assets.
 		if ( $data['release_asset'] ) {
-			unset( $data['branches'][ $data['release_branch'] ] );
+			unset( $data['branches'][ $data['production_branch'] ] );
 		}
 		if ( null !== $data['branches'] ) {
 			foreach ( array_keys( $data['branches'] ) as $branch ) {