+++ src/GitHub_Updater/Theme.php
@@ -190,14 +190,14 @@ class Theme {
 			$current_branch = "current_branch_{$header['repo']}";
 
 			if ( isset( self::$options[ $current_branch ] )
-			&& ( 'master' === self::$options[ $current_branch ] && 'master' !== $header['release_branch'] )
+			&& ( 'master' === self::$options[ $current_branch ] && 'master' !== $header['production_branch'] )
 			) {
 				unset( self::$options[ $current_branch ] );
 				update_site_option( 'github_updater', self::$options );
 			}
 			$branch = isset( self::$options[ $current_branch ] )
 				? self::$options[ $current_branch ]
-				: $header['release_branch'];
+				: $header['production_branch'];
 
 			$git_theme['type']                    = 'theme';
 			$git_theme['git']                     = $repo_parts['git_server'];
@@ -215,7 +215,7 @@ class Theme {
 			$git_theme['sections']['description'] = $theme['Description'];
 			$git_theme['local_path']              = get_theme_root() . '/' . $git_theme['slug'] . '/';
 			$git_theme['branch']                  = $branch;
-			$git_theme['release_branch']          = $header['release_branch'];
+			$git_theme['production_branch']       = $header['production_branch'];
 			$git_theme['languages']               = $header['languages'];
 			$git_theme['ci_job']                  = $header['ci_job'];
 			$git_theme['release_asset']           = $header['release_asset'];
@@ -456,7 +456,7 @@ class Theme {
 		$branch_switch_data['branch']            = $branch;
 		$branch_switch_data['branches']          = $branches;
 		$branch_switch_data['release_asset']     = $repo->release_asset;
-		$branch_switch_data['release_branch']    = $repo->release_branch;
+		$branch_switch_data['production_branch'] = $repo->production_branch;
 
 		/*
 		 * Create after_theme_row_
@@ -615,9 +615,9 @@ class Theme {
 				<option value=""><?php esc_html_e( 'Choose a Version', 'github-updater' ); ?>&#8230;</option>
 			<?php
 
-			// Disable branch switching to release branch for release assets.
+			// Disable branch switching to production branch for release assets.
 			if ( $theme->release_asset ) {
-				unset( $theme->branches[ $theme->release_branch ] );
+				unset( $theme->branches[ $theme->production_branch ] );
 			}
 			if ( isset( $theme->branches ) ) {
 				foreach ( array_keys( $theme->branches ) as $branch ) {
@@ -703,8 +703,8 @@ class Theme {
 					continue;
 				}
 
-				// Update download link for release_asset non-release branches.
-				if ( $theme->release_asset && $theme->release_branch !== $theme->branch ) {
+				// Update download link for release_asset non-production branches.
+				if ( $theme->release_asset && $theme->production_branch !== $theme->branch ) {
 					$response['package'] = $theme->branches[ $theme->branch ]['download'];
 				}